refactor: rename and move templates
