+++ backend/battling/views.py
@@ -2,42 +2,40 @@ from django.shortcuts import redirect, render
 from django.views.generic.base import TemplateView
 from django.views.generic.edit import CreateView
 
-from .battles.battle import battle
-from .battles.email import send_battle_result
-from .forms import CreatorForm, OpponentForm
-from .models import Battle
-from .pokemons.pokemon_data import get_pokemon_from_api, valid_team
+from battling.forms import CreatorForm, OpponentForm
+from battling.models import Battle
+from services.battles import battle
+from services.email import send_battle_result
+from services.pokemon_data import get_pokemon_from_api, valid_team
 
 
 class Home(TemplateView):
-    template_name = "home.html"
+    template_name = "battling/home.html"
 
 
 class Invite(TemplateView):
-    template_name = "invite.html"
+    template_name = "battling/invite.html"
 
 
 class Opponent(TemplateView):  # will become an email
-    template_name = "opponent.html"
+    template_name = "battling/opponent.html"
 
 
 class CreateBattle(CreateView):
     model = Battle
     form_class = CreatorForm
-    template_name = "create_battle.html"
+    template_name = "battling/create_battle.html"
 
-    def get_c(self, request):
-        return render(request, self.template_name, {"form": self.form_class})
-
-    def post_c(self, request):
+    def post(self, request, *args, **kwargs):
         battle_field = self.model.objects.latest("id")
         form = self.form_class(request.POST, instance=battle_field)
+        # This type of validation should be on the form CreatorForm
         if form.is_valid():
             round_battle = form.save(commit=False)
             c_team = valid_team(round_battle, "c")
             if c_team:
                 form.save()
-                return redirect("Invite")
+                return redirect("invite")
             message = "ERROR: you selected sum more than 600 points"
             return render(
                 request, self.template_name, {"form": self.form_class, "message": message}
@@ -46,14 +44,12 @@ class CreateBattle(CreateView):
 
 
 class EnterBattle(CreateView):
+    # The same comments made for EnterBattle apply to this one.
     model = Battle
     form_class = OpponentForm
-    template_name = "enter_battle.html"
+    template_name = "battling/enter_battle.html"
 
-    def get_o(self, request):
-        return render(request, self.template_name, {"form": self.form_class})
-
-    def post_o(self, request):
+    def post(self, request, *args, **kwargs):
         battle_field = self.model.objects.latest("id")
         form = self.form_class(request.POST, instance=battle_field)
         if form.is_valid():
@@ -88,12 +84,12 @@ def battle_end(request):
     return render(request, "battling/battle_end.html")
 
 
-def fights(request):
+def battle_details(request):
     battle_id = Battle.objects.latest("id")
     battle_info = Battle.objects.filter(id=battle_id.id).values()[0]
 
-    creator_pkms = [get_pokemon_from_api(battle_info["c_pokemon" + str(i)]) for i in range(1, 4)]
-    opponent_pkms = [get_pokemon_from_api(battle_info["o_pokemon" + str(i)]) for i in range(1, 4)]
+    creator_pkms = [get_pokemon_from_api(battle_info["c_pokemon_" + str(i)]) for i in range(1, 4)]
+    opponent_pkms = [get_pokemon_from_api(battle_info["o_pokemon_" + str(i)]) for i in range(1, 4)]
 
     score = battle(creator_pkms, opponent_pkms)
 
@@ -105,7 +101,7 @@ def fights(request):
 
     return render(
         request,
-        "battling/fights.html",
+        "battling/battle_details.html",
         {
             "winner": winner,
             "creator_pkms": creator_pkms,
