+++ backend/templates/battling/base.html
@@ -1,3 +1,9 @@
+<!-- Please move this an scss file
+Also, try to add colors to variables and use them on the files.
+$color-white-gray: #ffffff
+and
+background-color: $color-white-gray -->
+
 <!DOCTYPE html>
 <html lang="en">
