+++ backend/templates/battling/battle_end.html
