+++ backend/templates/battling/create_battle.html
