+++ backend/templates/battling/enter_battle.html
