+++ backend/templates/battling/home.html
@@ -1,7 +1,7 @@
 {% extends "battling/base.html" %}
 
 {% block content %}
-  <a class="button_battle"  href="{% url 'CreateBattle' %}" role="button">
+  <a class="button_battle"  href="{% url 'create_battle' %}" role="button">
     <p class="text_button">Create a new battle</p>
   </a>
 {% endblock %}
