+++ backend/templates/battling/invite.html
