+++ backend/templates/battling/opponent.html
@@ -3,7 +3,7 @@
 {% block content %}
   <div class="teste">
     <p class="invite">You've been invited to fight.</p>
-    <a class="button1" href="{% url 'EnterBattle' %}" role="button">
+    <a class="button1" href="{% url 'enter_battle' %}" role="button">
       <p class="text_button">Check it out!</p>
     </a>
   </div>
