+++ backend/templates/templated_email/battle_result.email
